feat: redirect from / to /arr
